+++ src/Carousel/Carousel.js
@@ -4,6 +4,9 @@ import ChevronLeftLarge from 'wix-ui-icons-common/ChevronLeftLarge';
 import ChevronRightLarge from 'wix-ui-icons-common/ChevronRightLarge';
 import ChevronLeftLargeSmall from 'wix-ui-icons-common/ChevronLeftLargeSmall';
 import ChevronRightLargeSmall from 'wix-ui-icons-common/ChevronRightLargeSmall';
+import ChevronLeftSmall from 'wix-ui-icons-common/ChevronLeftSmall';
+import ChevronRightSmall from 'wix-ui-icons-common/ChevronRightSmall';
+
 // This is here and not in the test setup because we don't want consumers to need to run it as well
 import '../common/match-media-register';
 import Slider from 'react-slick';
@@ -49,6 +52,9 @@ class Carousel extends React.Component {
     /** Sets the skin of the arrow buttons */
     buttonSkin: PropTypes.oneOf(['standard', 'inverted', 'light']),
 
+    /** Show a shadow for the carousel controls */
+    showControlsShadow: PropTypes.bool,
+
     /** Images loop endlessly */
     infinite: PropTypes.bool,
 
@@ -75,6 +81,9 @@ class Carousel extends React.Component {
 
     /** Sets the arrows position */
     controlsSize: PropTypes.oneOf(['tiny', 'small', 'medium']),
+
+    /** Configure the start and end controls to be shown disabled or hidden. Relevant when infinite prop is set to false. */
+    controlsStartEnd: PropTypes.oneOf(['disabled', 'hidden']),
   };
 
   static defaultProps = {
@@ -86,6 +95,8 @@ class Carousel extends React.Component {
     buttonSkin: 'standard',
     controlsPosition: 'sides',
     controlsSize: 'medium',
+    controlsStartEnd: 'disabled',
+    showControlsShadow: false,
   };
 
   constructor(props) {
@@ -104,6 +115,7 @@ class Carousel extends React.Component {
       children,
       controlsPosition,
       controlsSize,
+      showControlsShadow,
     } = this.props;
     const { sliderSettings } = this.state;
     const hasImages = !children && images.length > 0;
@@ -113,7 +125,7 @@ class Carousel extends React.Component {
         data-hook={dataHook}
         className={st(
           classes.root,
-          { controlsPosition, controlsSize },
+          { controlsPosition, controlsSize, showControlsShadow },
           className,
         )}
       >
@@ -124,6 +136,28 @@ class Carousel extends React.Component {
       </div>
     );
   }
+
+  leftIconByControlSize = controlSize => {
+    switch (controlSize) {
+      case 'tiny':
+        return <ChevronLeftSmall />;
+      case 'small':
+        return <ChevronLeftLargeSmall />;
+      default:
+        return <ChevronLeftLarge />;
+    }
+  };
+
+  rightIconByControlSize = controlSize => {
+    switch (controlSize) {
+      case 'tiny':
+        return <ChevronRightSmall />;
+      case 'small':
+        return <ChevronRightLargeSmall />;
+      default:
+        return <ChevronRightLarge />;
+    }
+  };
 
   _resolveSliderSettings = ({
     infinite,
@@ -136,6 +170,7 @@ class Carousel extends React.Component {
     beforeChange,
     controlsPosition,
     controlsSize,
+    controlsStartEnd,
   }) => {
     return {
       infinite,
@@ -154,13 +189,8 @@ class Carousel extends React.Component {
           dataHook={dataHooks.nextButton}
           buttonSkin={buttonSkin}
           arrowSize={controlsSize}
-          icon={
-            controlsSize === 'tiny' || controlsSize === 'small' ? (
-              <ChevronRightLargeSmall />
-            ) : (
-              <ChevronRightLarge />
-            )
-          }
+          icon={this.rightIconByControlSize(controlsSize)}
+          controlsStartEnd={controlsStartEnd}
         />
       ),
       prevArrow: (
@@ -168,13 +198,8 @@ class Carousel extends React.Component {
           dataHook={dataHooks.prevButton}
           buttonSkin={buttonSkin}
           arrowSize={controlsSize}
-          icon={
-            controlsSize === 'tiny' || controlsSize === 'small' ? (
-              <ChevronLeftLargeSmall />
-            ) : (
-              <ChevronLeftLarge />
-            )
-          }
+          icon={this.leftIconByControlSize(controlsSize)}
+          controlsStartEnd={controlsStartEnd}
         />
       ),
       arrows: controlsPosition !== 'none',