+++ src/Carousel/docs/examples.js
@@ -29,6 +29,27 @@ export const controlsPositionExample = `
 />
 `;
 
+export const controlsStartEndHidden = `
+<Carousel
+  buttonSkin="inverted"
+  controlsPosition="overlay"
+  controlsSize="medium"
+  infinite={false}
+  controlsStartEnd="hidden"
+  images={[
+    {
+      src: 'https://a-static.besthdwallpaper.com/garfield-wallpaper-2800x2100-815_28.jpg'
+    },
+    {
+      src: 'https://m.media-amazon.com/images/M/MV5BZGMwOGIwZjUtOWQ1OS00YWRjLWJmZGMtN2Y1OWQ3ZDYwYTM3XkEyXkFqcGdeQXVyNzU1NzE3NTg@._V1_.jpg'
+    },
+    {
+      src: 'https://a-static.besthdwallpaper.com/cartoons-garfield-wallpaper-1440x1080-6773_22.jpg'
+    }
+  ]}
+/>
+`;
+
 export const autoplayExample = `
 <Carousel
   autoplay
@@ -49,6 +70,26 @@ export const autoplayExample = `
 export const withoutDotsExample = `
 <Carousel
   dots={false}
+  images={[
+    {
+      src: 'https://a-static.besthdwallpaper.com/garfield-wallpaper-2800x2100-815_28.jpg'
+    },
+    {
+      src: 'https://m.media-amazon.com/images/M/MV5BZGMwOGIwZjUtOWQ1OS00YWRjLWJmZGMtN2Y1OWQ3ZDYwYTM3XkEyXkFqcGdeQXVyNzU1NzE3NTg@._V1_.jpg'
+    },
+    {
+      src: 'https://a-static.besthdwallpaper.com/cartoons-garfield-wallpaper-1440x1080-6773_22.jpg'
+    }
+  ]}
+/>
+`;
+
+export const controlsShadowExample = `
+<Carousel
+  buttonSkin="inverted"
+  controlsPosition="overlay"
+  controlsSize="medium"
+  showControlsShadow
   images={[
     {
       src: 'https://a-static.besthdwallpaper.com/garfield-wallpaper-2800x2100-815_28.jpg'