+++ src/Carousel/test/Carousel.tsx
@@ -25,6 +25,8 @@ function CarouselWithAllProps() {
       initialSlideIndex={1}
       variableWidth
       controlsPosition="sides"
+      controlsStartEnd="disabled"
+      showControlsShadow={false}
     />
   );
 }