+++ src/SidebarSectionItem/test/SidebarSectionItem.visual.js
@@ -5,6 +5,7 @@ import Sound from 'wix-ui-icons-common/Sound';
 import SidebarSectionItem from '../SidebarSectionItem';
 import { SidebarContext } from '../../Sidebar/SidebarAPI';
 import Box from '../../Box';
+import Badge from '../../Badge';
 
 const skins = ['dark', 'light'];
 
@@ -17,6 +18,12 @@ const SamplePrefix = () => (
 );
 
 const SampleSuffix = () => <Sound />;
+
+const SampleNewBadgeSuffix = () => (
+  <Badge size="small" skin="warning" type="solid" uppercase>
+    NEW
+  </Badge>
+);
 
 const tests = [
   {
@@ -136,6 +143,14 @@ const tests = [
           suffix: <SampleSuffix />,
         },
       },
+      {
+        it: 'Should display suffix with new badge and multiple lines',
+        props: {
+          children:
+            'This is a very long text which exceeds the maximum width of its container',
+          suffix: <SampleNewBadgeSuffix />,
+        },
+      },
     ],
   },
 ];
@@ -166,5 +181,3 @@ export const runTests = (
     ),
   );
 };
-
-runTests();