+++ src/index.js
@@ -45,7 +45,10 @@ export { default as Breadcrumbs } from './Breadcrumbs';
 export { default as ImageViewer } from './ImageViewer';
 export { default as Badge } from './Badge';
 export { default as BadgeSelect } from './BadgeSelect';
-export { badgeSelectItemBuilder } from './BadgeSelectItemBuilder';
+export {
+  default as BadgeSelectItem,
+  badgeSelectItemBuilder,
+} from './BadgeSelectItem';
 export { contactItemBuilder } from './ContactItemBuilder';
 export { default as EditableSelector } from './EditableSelector';
 export { default as Card } from './Card';
@@ -189,3 +192,5 @@ export { default as AddressInput } from './AddressInput';
 export { default as StackedBarChart } from './StackedBarChart';
 export { default as CardFolderTabs } from './CardFolderTabs';
 export { default as PageFooter } from './PageFooter';
+export { default as StyledNestableList } from './StyledNestableList';
+export { default as TrendIndicator } from './TrendIndicator';